feat(buyer-Info): abstract details through hooks
